+++ minn-admin.php
@@ -32,6 +32,7 @@ require_once MINN_ADMIN_DIR . 'includes/adapters/gravity-forms.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/fluent-forms.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/ninja-forms.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/forminator.php';
+require_once MINN_ADMIN_DIR . 'includes/adapters/formidable.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/cf7-flamingo.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/cfdb7.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/elementor-forms.php';